refactor(cli): decode the code-mode API with the server's wire types
The CLI kept an 850-line mirror of the code-mode snapshots, the sequenced
event frame, and the /code/updates notices, and it had drifted from the
server. Re-export those types from tidebreak_server::wire, derive
Deserialize and deny_unknown_fields on them, and import them in the CLI.

Write one real value per snapshot kind, notice variant, and event variant
to crates/tidebreak-server/fixtures/code-frames.json from the server's own
types, and decode every entry from the server, the CLI, and the desktop's
code parsers. The desktop pass found that the socket parser dropped frames
carrying transient, replacement, or truncated flags and that no arm read
the delivery notice; both are fixed and pinned. Sessions that bind no
workspace still fail the desktop parsers and are listed as known gaps
under #3027.

Closes #3022

+++ crates/tidebreak-cli/src/agent_mcp/code_follow.rs
@@ -255,7 +255,7 @@ async fn reconcile_lost_socket(
             pending: None,
             events_cursor,
             turn_id: Some(turn_id),
-            queue_position: Some(row.position),
+            queue_position: Some(i64::from(row.position)),
         });
     }
     Ok(running(assistant_text, events_cursor, Some(turn_id)))
@@ -363,7 +363,7 @@ pub(crate) async fn run_turn(
                             pending: None,
                             events_cursor: stream.last_seq(),
                             turn_id: Some(row.id),
-                            queue_position: Some(row.position),
+                            queue_position: Some(i64::from(row.position)),
                         });
                     }
                 }
